third_party/WebKit/LayoutTests/inspector/elements/breadcrumb-shadow-roots.html

<!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN" "http://www.w3.org/TR/html4/loose.dtd">
<html>
<head>
<script src="../../http/tests/inspector/inspector-test.js"></script>
<script src="../../http/tests/inspector/elements-test.js"></script>
<script>

function loaded()
{
    var template = document.querySelector("#tmpl");
    var root = document.querySelector("#host").createShadowRoot();
    root.appendChild(template.content.cloneNode(true));
    runTest();
}

function test()
{
    WebInspector.inspectorView._showPanel("elements");
    WebInspector.settings.showUAShadowDOM.set(true);
    InspectorTest.expandElementsTree(step0);

    function step0()
    {
        selectNode(matchUserAgentShadowRoot, step1);
    }

    function step1()
    {
        InspectorTest.dumpBreadcrumb("User-agent shadow root breadcrumb");
        selectNode(matchAuthorShadowRoot, step2);
    }

    function step2()
    {
        InspectorTest.dumpBreadcrumb("Author shadow root breadcrumb");
        InspectorTest.completeTest();
    }

    function selectNode(matchFunction, next)
    {
        InspectorTest.findNode(matchFunction, callback);
        function callback(node)
        {
            WebInspector.Revealer.revealPromise(node).then(next);
        }
    }

    function matchUserAgentShadowRoot(node)
    {
        return node.shadowRootType() === WebInspector.DOMNode.ShadowRootTypes.UserAgent;
    }

    function matchAuthorShadowRoot(node)
    {
        return node.shadowRootType() === WebInspector.DOMNode.ShadowRootTypes.Author;
    }
}

</script>
</head>

<body onload="loaded()">
<p>
Tests that shadow roots are displayed correctly in breadcrumbs.
</p>

<input type="text">
<div id="host"></div>
<template id="tmpl">
    <style>.red { color: red; }</style>
    <div id="inner" class="red">inner</div>
</template>


</body>
</html>
