third_party/WebKit/LayoutTests/inspector/elements/breadcrumb-shadow-roots-expected.txt

Tests that shadow roots are displayed correctly in breadcrumbs.


User-agent shadow root breadcrumb:
html > body > input > #shadow-root
Author shadow root breadcrumb:
html > body > div#host > #shadow-root
